Normalize Unit Testing workflow across repos (#571)
* Revert "Apply fixes from StyleCI"

This reverts commit c7b6de34cb0280f0d3374ae6587438ae1d693f96.

* Better group name for tests that requires DNS

* Revert "Better group name for tests that requires DNS"

This reverts commit 12274e989f90f8c050574d94a7eb730ccd5535b3.

* do not install unneeded packages

* better upload coverage

* shorter title

* Revert "shorter title"

This reverts commit edf6c16077dad0688a87af3d2cbd6ab452b11542.

* sort composer packages

* Add DarkSide666 to authors in composer

* add lint test

* Normalize Unit Testing workflow across repos

* Fix PSR

* normalize DB config

* norm .gitignore

* Add CS fixer rules

* fix install like in core

* apply safe/mass CS fixes

* apply risky/mass CS fixes

* manual/checked CS fixes

* verbose phpunit

* improve location of csv test files

* more strict CS

* strict comparison

* assertSame

* Update composer.json

* style fix

* oh

+++ tests/CSVTest.php
@@ -12,34 +12,46 @@ use atk4\data\tests\Model\Person as Person;
  */
 class CSVTest extends AtkPhpunit\TestCase
 {
-    public $file = 'atk-test.csv';
-    public $file2 = 'atk-test-2.csv';
+    protected $file;
+    protected $file2;
 
-    public function setDB($data)
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        // better to skip this test on Windows, prevent permissions issues
+        // see also https://github.com/atk4/data/issues/271
+        if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+            $this->markTestSkipped('Skip on Windows');
+        }
+
+        $this->file = sys_get_temp_dir() . '/atk4_test__data__a.csv';
+        $this->file2 = sys_get_temp_dir() . '/atk4_test__data__b.csv';
+    }
+
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+
+        if (file_exists($this->file)) {
+            unlink($this->file);
+        }
+        if (file_exists($this->file2)) {
+            unlink($this->file2);
+        }
+    }
+
+    protected function setDB($data): void
     {
         $f = fopen($this->file, 'w');
-        fputcsv($f, array_keys(current($data)));
+        fputcsv($f, array_keys(reset($data)));
         foreach ($data as $row) {
             fputcsv($f, $row);
         }
         fclose($f);
     }
 
-    public function tearDown(): void
-    {
-        parent::tearDown();
-
-        // see: https://github.com/atk4/data/issues/271
-        try {
-            unlink($this->file);
-            if (file_exists($this->file2)) {
-                unlink($this->file2);
-            }
-        } catch (\Exception $e) {
-        }
-    }
-
-    public function getDB()
+    protected function getDB(): array
     {
         $f = fopen($this->file, 'r');
         $keys = fgetcsv($f);
@@ -64,7 +76,7 @@ class CSVTest extends AtkPhpunit\TestCase
 
         $this->setDB($data);
         $data2 = $this->getDB();
-        $this->assertEquals($data, $data2);
+        $this->assertSame($data, $data2);
     }
 
     public function testLoadAny()
@@ -82,8 +94,8 @@ class CSVTest extends AtkPhpunit\TestCase
         $m->addField('surname');
         $m->loadAny();
 
-        $this->assertEquals('John', $m['name']);
-        $this->assertEquals('Smith', $m['surname']);
+        $this->assertSame('John', $m['name']);
+        $this->assertSame('Smith', $m['surname']);
     }
 
     public function testLoadAnyException()
@@ -102,8 +114,8 @@ class CSVTest extends AtkPhpunit\TestCase
         $m->loadAny();
         $m->loadAny();
 
-        $this->assertEquals('Sarah', $m['name']);
-        $this->assertEquals('Jones', $m['surname']);
+        $this->assertSame('Sarah', $m['name']);
+        $this->assertSame('Jones', $m['surname']);
 
         $m->tryLoadAny();
         $this->assertFalse($m->loaded());
@@ -112,8 +124,8 @@ class CSVTest extends AtkPhpunit\TestCase
     public function testPersistenceCopy()
     {
         $data = [
-            ['name' => 'John', 'surname' => 'Smith', 'gender'=>'M'],
-            ['name' => 'Sarah', 'surname' => 'Jones', 'gender'=>'F'],
+            ['name' => 'John', 'surname' => 'Smith', 'gender' => 'M'],
+            ['name' => 'Sarah', 'surname' => 'Jones', 'gender' => 'F'],
         ];
 
         $this->setDB($data);
@@ -129,7 +141,7 @@ class CSVTest extends AtkPhpunit\TestCase
             $m2->save($m);
         }
 
-        $this->assertEquals(
+        $this->assertSame(
             file_get_contents($this->file2),
             file_get_contents($this->file)
         );
@@ -151,12 +163,12 @@ class CSVTest extends AtkPhpunit\TestCase
         $m->addField('name');
         $m->addField('surname');
 
-        $this->assertEquals([
+        $this->assertSame([
             ['id' => 1, 'name' => 'John', 'surname' => 'Smith'],
             ['id' => 2, 'name' => 'Sarah', 'surname' => 'Jones'],
         ], $m->export());
 
-        $this->assertEquals([
+        $this->assertSame([
             ['surname' => 'Smith'],
             ['surname' => 'Jones'],
         ], $m->export(['surname']));
